refactor : 패키지 구조 수정 - 최상위 network package 안에 repository, service package 이동 - 그로 인한 Package, import 구문 수정
