+++ app/src/main/java/com/sangik/iluvbook/fairytale/creation/ui/CustomChipView.kt
@@ -1,4 +1,4 @@
-package com.sangik.iluvbook.fairytalecreation.ui
+package com.sangik.iluvbook.fairytale.creation.ui
 
 import android.content.Context
 import android.util.AttributeSet
