+++ app/src/main/java/com/sangik/iluvbook/network/service/HangmanService.kt
@@ -1,5 +1,6 @@
-package com.sangik.iluvbook.hangman.game.model
+package com.sangik.iluvbook.network.service
 
+import com.sangik.iluvbook.hangman.game.model.HangmanResponse
 import retrofit2.Response
 import retrofit2.http.GET
